Responded to review + simplified tests quite a bit (removed autodiff tests for append_array) (Issue 481)
